+++ fheroes2.vcxproj
@@ -518,6 +518,7 @@
     <ClInclude Include="src\fheroes2\system\settings.h" />
     <ClInclude Include="src\fheroes2\world\world.h" />
     <ClInclude Include="src\fheroes2\world\world_pathfinding.h" />
+    <ClInclude Include="src\fheroes2\world\world_regions.h" />
     <ClInclude Include="src\thirdparty\libsmacker\smacker.h" />
     <ClInclude Include="src\thirdparty\libsmacker\smk_bitstream.h" />
     <ClInclude Include="src\thirdparty\libsmacker\smk_hufftree.h" />
